Add files via upload
These are two queries, one that I used as the basis for a VLOOKUP sorting units into different labor force categories, this was necessary as the stats needed to break up the different groups are contained in different variables. So we'll have to use some IF statements to divide them up. 

Unit income is defined as household income divided by the square root of the number of individuals in the household.